+++ src/main/java/ch/uzh/ifi/seal/soprafs20/rest/mapper/DTOMapper.java
@@ -127,7 +127,7 @@ public interface DTOMapper {
     @Mapping(source = "wonCards", target = "wonCards")
     @Mapping(source = "lostCards", target = "lostCards")
     @Mapping(source = "mysteryWord", target = "mysteryWord")
-    GuessGetDTO convertEntityToGuessGETDTO(String guess, boolean success,int leftCards, Long wonCards, int lostCards, String mysteryWord);
+    GuessGetDTO convertEntityToGuessGETDTO(String guess, boolean success,int leftCards, int wonCards, int lostCards, String mysteryWord);
 
     @Named("convertString")
     default String convertGuessPostDTOToGuessString(GuessPostDTO guessPostDTO)
